reorder lib files
