+++ include/glap/common/utils.h
@@ -1,7 +1,7 @@
 #pragma once 
 #include <concepts>
-#include "expected.h"
-#include "utf8.h"
+#include "../common/expected.h"
+#include "../common/utf8.h"
 
 namespace glap
 {
