+++ include/glap/v1/help_formatter.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <fmt/format.h>
+#include "fmt/core.h"
+#include "config.h"
+namespace glap 
+{
+    inline auto default_help_formatter() {
+        return [](const config::Command& command) {
+            fmt::print("{}", command.longname);
+        };
+    }
+}
