+++ include/glap/v1/parser.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "utils.h"
+#include "../common/utils.h"
+#include "../common/error.h"
 #include "config.h"
 #include "result.h"
 
@@ -26,7 +27,7 @@ namespace glap
             global_command = longname;
             return *this;
         }
-        auto parse(utils::Iterable<std::string_view> auto args) const -> result::PosExpected<result::Result>;
+        auto parse(utils::Iterable<std::string_view> auto args) const -> PosExpected<result::Result>;
     private:
         auto get_global_command() const -> std::optional<std::reference_wrapper<const config::Command>> {
             if (!global_command.has_value()) 
@@ -43,15 +44,15 @@ namespace glap
             return std::nullopt;
         }
         template <utils::Iterator<std::string_view> Iter>
-        auto parse_long_argument(Iter& itarg, Iter end, const config::Command& command, result::Command& result_command) const -> result::PosExpected<bool>;
+        auto parse_long_argument(Iter& itarg, Iter end, const config::Command& command, result::Command& result_command) const -> PosExpected<bool>;
         template <utils::Iterator<std::string_view> Iter>
-        auto parse_short_argument(Iter& itarg, Iter end, const config::Command& command, result::Command& result_command) const -> result::PosExpected<bool>;
+        auto parse_short_argument(Iter& itarg, Iter end, const config::Command& command, result::Command& result_command) const -> PosExpected<bool>;
         
-        auto add_argument(result::Command& result_command, const config::Argument& argument, std::string_view name, const std::string_view value) const -> result::PosExpected<bool>;
-        auto add_flag(result::Command& result_command, const config::Flag& flag, std::string_view name) const -> result::PosExpected<bool>;
-        auto add_input(result::Command& result_command, const config::Command& command, std::string_view input) const -> result::PosExpected<bool>;
+        auto add_argument(result::Command& result_command, const config::Argument& argument, std::string_view name, const std::string_view value) const -> PosExpected<bool>;
+        auto add_flag(result::Command& result_command, const config::Flag& flag, std::string_view name) const -> PosExpected<bool>;
+        auto add_input(result::Command& result_command, const config::Command& command, std::string_view input) const -> PosExpected<bool>;
 
-        auto parse_command(utils::Iterable<std::string_view> auto args, const config::Command& command) const -> result::PosExpected<result::Command>;
+        auto parse_command(utils::Iterable<std::string_view> auto args, const config::Command& command) const -> PosExpected<result::Command>;
 
     };
 }
