+++ include/glap/v2/parser.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "expected.h"
+#include "../common/expected.h"
 #include <algorithm>
 #include <cstddef>
 #include <limits>
